+++ elfo-logger/src/filtering_layer.rs
@@ -1,7 +1,7 @@
 use tracing::{subscriber::Interest, Level, Metadata, Subscriber};
 use tracing_subscriber::layer::{Context, Layer};
 
-use elfo_core::scope;
+use elfo_core::{logging::_priv::CheckResult, scope};
 
 use crate::stats;
 
@@ -17,27 +17,29 @@ impl FilteringLayer {
 }
 
 impl<S: Subscriber> Layer<S> for FilteringLayer {
-    fn register_callsite(&self, _metadata: &'static Metadata<'static>) -> Interest {
+    fn register_callsite(&self, _meta: &'static Metadata<'static>) -> Interest {
         Interest::sometimes()
     }
 
-    fn enabled(&self, metadata: &Metadata<'_>, _ctx: Context<'_, S>) -> bool {
+    fn enabled(&self, meta: &Metadata<'_>, cx: Context<'_, S>) -> bool {
         scope::try_with(|scope| {
-            let level = *metadata.level();
+            let level = *meta.level();
 
             if !scope.permissions().is_logging_enabled(level) {
                 return false;
             }
 
-            if scope.logging_limiter().acquire() {
-                true
-            } else {
-                stats::counter_per_level("elfo_limited_events_total", level);
-                false
+            match scope.logging().check(meta, cx) {
+                CheckResult::Passed => true,
+                CheckResult::NotInterested => false,
+                CheckResult::Limited => {
+                    stats::counter_per_level("elfo_limited_events_total", level);
+                    false
+                }
             }
         })
         // TODO: limit events outside the actor system?
-        .unwrap_or_else(|| *metadata.level() <= self.other_max_level)
+        .unwrap_or_else(|| *meta.level() <= self.other_max_level)
     }
 
     // TODO: implement `max_level_hint()`